[review service] refactored domain layer code
