+++ review/src/main/java/com/open/forum/review/domain/model/comment/CommentContent.java
@@ -5,6 +5,8 @@ import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
+import java.time.ZonedDateTime;
+
 /**
  * Represents the content of a comment in the review system.
  * This class contains information about the comment's ID and its content.
@@ -16,4 +18,6 @@ public class CommentContent {
     @NotEmpty
     @Max(value = 5000, message = "Comment must be less than 5000 characters")
     private final String content;
+    @NotNull
+    private final ZonedDateTime createdAt;
 }
